# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "SuperMakers"
        }
    ],
    "paths": {
...
        "/api/payment-statuses": {
            "get": {
                "tags": [
                    "Payment Statuses"
                ],
                "summary": "Get a list of payment statuses",
                "description": "Returns a list of all payment statuses.",
                "operationId": "3b5bf463bebe2112722517b12cd5627d",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Payment Statuses"
                ],
                "summary": "Create a new payment status",
                "description": "Adds a new payment status to the database",
                "operationId": "6206ca1f77d8143062d8ef7bbe2e029b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Paid"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Payment has been completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment status created"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/payment-statuses/{id}": {
            "get": {
                "tags": [
                    "Payment Statuses"
                ],
                "summary": "Get a payment status by ID",
                "description": "Returns a single payment status",
                "operationId": "17ee5abdc5991e8b1b60bd588e71c073",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of payment status to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment status found"
                    },
                    "404": {
                        "description": "Payment status not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Payment Statuses"
                ],
                "summary": "Update an existing payment status",
                "description": "Updates payment status details by ID",
                "operationId": "22d9278c7528f6943bc4919d7eacf0f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of payment status to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Paid Updated"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Payment has been completed and verified"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment status updated"
                    },
                    "404": {
                        "description": "Payment status not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payment Statuses"
                ],
                "summary": "Delete a payment status",
                "description": "Deletes a payment status by ID",
                "operationId": "d9e4d3f4faeab3b4e1af8dd4066a9dce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of payment status to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Payment status deleted"
                    },
                    "404": {
                        "description": "Payment status not found"
                    }
                }
            }
        },
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get list of roles",
                "operationId": "ca86f0a206144dfbd695b749b4b28f9d",
                "responses": {
                    "200": {
                        "description": "List of roles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create a new role",
                "operationId": "341b8b162b61267c7bd187500aea5bc4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role created successfully"
                    },
                    "400": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get role by ID",
                "operationId": "15c404caf238f1861c81f68205d328e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role data"
                    },
                    "404": {
                        "description": "Role not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Update role",
                "operationId": "c73ab816ae8fb7c34a995d37ba577082",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role updated successfully"
                    },
                    "404": {
                        "description": "Role not found"
                    },
                    "400": {
                        "description": "Validation errors"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Delete role",
                "operationId": "ffa6ef115d86b00cc6c54a6210019f9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role deleted successfully"
                    },
                    "404": {
                        "description": "Role not found"
                    }
                }
            }
        },
...
    },
...
}